safer clear webpack
disallows deletion of the current project.
disallows deletion of directories outside current project.

+++ index.js
@@ -22,10 +22,18 @@ Plugin.prototype.apply = function(compiler) {
   if(self.paths === void 0)
     return
 
+  var projectDir = path.dirname(module.parent.filename)
+
   // preform an rm -rf on each path
   self.paths.forEach(function(path){
     path = resolve(self.context, path);
-    rimraf.sync(path);
+
+    // disallow deletion of project path and any directories outside of project path.
+    if(path != projectDir && path.indexOf(projectDir) === 0){
+      rimraf.sync(path);
+    } else {
+      console.warn("Clean webpack did not delete path: " + path);
+    }
   });
 };
